BACKEND-Filter stops: Move functions to right files and new class
