+++ backend/trips.py
@@ -1,5 +1,6 @@
 from backend.connect_to_api import ResRobot
 import pandas as pd
+import re
 
 resrobot = ResRobot()
 
@@ -83,6 +84,77 @@ class TripPlanner:
         return []
 
 
+class StopPlanner():
+
+    def __init__(self, station) -> None:
+        self.station = station
+
+        # Filter stops on arrival and departure v
+
+    def get_timetable_dep(self, station):
+        if station:
+            station_name = station[0]["StopLocation"]["name"]
+            station_id_raw = station[0]["StopLocation"]["id"]
+
+            match = re.search(r"L=(\d+)", station_id_raw)
+            if match:
+                station_id = match.group(1)
+            else:
+                print("Kunde inte extrahera stationens ID.")
+                station_id = None
+            print(f"Stationens namn: {station_name}")
+            print(f"Stations id: {station_id}")
+
+            departures = resrobot.get_departures(station_id, max_results=8)
+
+            if departures:
+                dep_data = []
+                for departure in departures:
+                    transport = departure.get('ProductAtStop', {}).get(
+                        'displayNumber', 'Okänt fordon')
+                    dep_data.append(
+                        {"Tid": departure['time'], "Destination": departure['direction'], "Linje": transport})
+                return pd.DataFrame(dep_data)
+            else:
+                print("Inga avgångar")
+                return pd.DataFrame()
+        else:
+            print("Stationen hittades inte")
+            return pd.DataFrame()
+
+    def get_timetable_arr(self, station):
+        if station:
+            station_name = station[0]["StopLocation"]["name"]
+            station_id_raw = station[0]["StopLocation"]["id"]
+
+            match = re.search(r"L=(\d+)", station_id_raw)
+            if match:
+                station_id = match.group(1)
+            else:
+                print("Kunde inte extrahera stationens ID.")
+                station_id = None
+            print(f"Stationens namn: {station_name}")
+            print(f"Stations id: {station_id}")
+
+            arrivals = resrobot.get_arrivials(station_id, max_results=8)
+
+            if arrivals:
+                arr_data = []
+                for arrival in arrivals:
+                    transport = arrival.get('ProductAtStop', {}).get(
+                        'displayNumber', 'Okänt fordon')
+                    arr_data.append({
+                        "Tid": arrival['time'], "Origin": arrival['origin'], "Linje": transport})
+                return pd.DataFrame(arr_data)
+            else:
+                print("Inga avgångar")
+                return pd.DataFrame()
+        else:
+            print("Stationen hittades inte")
+            return pd.DataFrame()
+    # Filter stops on arrival and departure ^
+
+
 if __name__ == "__main__":
-    data = TripPlanner(740000190, 740000001)
-    print(data.next_available_trip()[["arrTime", "depTime", "time", "date"]])
+    TripPlanner(740000190, 740000001)
+    StopPlanner()
